refactor(debates): move stance cards to SingleDebate, remove Top Comments from DebateScreen
- DebateScreen: show only main question card + Join the conversation; remove Agree/Disagree/Wildcard cards and Top Comments section
- SingleDebateScreen: add Agree/Disagree/Wildcard cards above Top Comments with vote bars and Vote Now links (mock voting)
- DebateScreen: drop comment state, mock comments, and related styles

Made-with: Cursor
